refactor(inbound): make the result state the single owner of visibility (#854)
Folds a Codex code-quality review's BLOCKING finding on 39a7f497. The review
judged that `InboundSubmissionResult` only *resembled* the
`RunFailureDiagnostic` pattern it was meant to follow, because it could still
express a contradiction:

- `owned_by_another_connection` repeated what `state` already said. One fact
  in two fields, free to disagree.
- `payload={}` was a placeholder standing in for "no payload". `{}` means an
  empty payload, and nothing stopped a NOT_FOUND result carrying a full one.

Now the state owns visibility outright:

- the boolean is gone from the dataclass;
- `payload` is `dict | None`, defaulting to None;
- `__post_init__` enforces the biconditional — a payload is present if and
  only if the state is FOUND — so the contradictory combinations are no
  longer constructible.

The wire response is deliberately unchanged: `_tool_get_result` still returns
`owned_by_another_connection: True` for the not-visible case, now derived as a
literal inside that branch. The key is useful to an MCP caller; only the
internal field was redundant.

Also folds the review's three non-blocking notes:

- `_result_api()` — a helper returning two imports as an untyped positional
  tuple — is deleted in favour of ordinary imports;
- the four visibility tests move out of the *reconciliation* suite into a new
  `tests/test_inbound_results.py`, importing the seeding helpers rather than
  copying them;
- the moved tests now assert `payload is None` on the unavailable states,
  pinning the new invariant.
